+++ packages/preset-typography/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pandacss/preset-typography
 
+## 2.0.0-beta.14
+
+### Patch Changes
+
+- @pandacss/types@2.0.0-beta.14
+
 ## 2.0.0-beta.13
 
 ### Patch Changes